No file reorder so tags make more sense

+++ profile.sh
@@ -17,7 +17,7 @@ cd "$tmpdir"
 
 ALTERED_PATH_FILES=""
 for f in ${INPUT_FILES[@]}; do
-    ALTERED_PATH_FILES="../$f $ALTERED_PATH_FILES"
+    ALTERED_PATH_FILES="$ALTERED_PATH_FILES ../$f"
 done
 
 clang-17 -gdwarf-4 -O0 -fpass-plugin=`echo ../keypoints/build/keypoints/KeyPointsPass.*` ${ALTERED_PATH_FILES[@]} ../keypoints/support/branchlog.c -o "$EXECUTABLE_NAME"
@@ -27,7 +27,11 @@ clang-17 -gdwarf-4 -O0 -fpass-plugin=`echo ../keypoints/build/keypoints/KeyPoint
 cd ..
 
 cp "$tmpdir"/branch_dictionary.txt ./branch_dictionary.txt."$pid"
-cp "$tmpdir"/branch_trace.txt ./branch_trace.txt."$pid"
+if [[ -f "$tmpdir"/branch_trace.txt ]]; then
+    cp "$tmpdir"/branch_trace.txt ./branch_trace.txt."$pid"
+else
+    touch ./branch_trace.txt."$pid"
+fi
 
 rm -rf "$tmpdir"
